requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.14
    # via langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via aiohttp
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
coloredlogs==15.0.1
    # via redisvl
dataclasses-json==0.6.7
    # via langchain-community
decorator==5.2.1
    # via ipython
distro==1.9.0
    # via openai
executing==2.2.0
    # via stack-data
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via langchain-community
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ipython==9.0.2
    # via -r requirements.in
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via langchain-redis
jiter==0.9.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.21
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.3.20
    # via -r requirements.in
langchain-core==0.3.47
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-redis
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-openai==0.3.9
    # via -r requirements.in
langchain-redis==0.2.0
    # via -r requirements.in
langchain-text-splitters==0.3.7
    # via langchain
langgraph==0.3.18
    # via -r requirements.in
langgraph-checkpoint==2.0.21
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==0.1.4
    # via langgraph
langgraph-sdk==0.1.58
    # via langgraph
langsmith==0.3.18
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
matplotlib-inline==0.1.7
    # via ipython
ml-dtypes==0.4.1
    # via redisvl
msgpack==1.1.0
    # via langgraph-checkpoint
multidict==6.2.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
neo4j==5.28.1
    # via -r requirements.in
numpy==1.26.4
    # via
    #   -r requirements.in
    #   langchain-community
    #   langchain-redis
    #   ml-dtypes
    #   redisvl
openai==1.68.2
    # via langchain-openai
orjson==3.10.15
    # via
    #   langgraph-sdk
    #   langsmith
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
prompt-toolkit==3.0.50
    # via ipython
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pydantic==2.10.6
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
    #   redisvl
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.8.1
    # via langchain-community
pygments==2.19.1
    # via
    #   ipython
    #   ipython-pygments-lexers
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-ulid==3.0.0
    # via
    #   langchain-redis
    #   redisvl
pytz==2025.1
    # via
    #   -r requirements.in
    #   neo4j
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   redisvl
redis==5.2.1
    # via
    #   -r requirements.in
    #   redisvl
redisvl==0.4.1
    # via langchain-redis
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.39
    # via
    #   langchain
    #   langchain-community
stack-data==0.6.3
    # via ipython
tabulate==0.9.0
    # via redisvl
tenacity==9.0.0
    # via
    #   langchain-community
    #   langchain-core
    #   redisvl
tiktoken==0.9.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via
    #   anyio
    #   ipython
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.3.0
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
yarl==1.18.3
    # via aiohttp
zstandard==0.23.0
    # via langsmith
